+++ SS14.Shared/Network/NetChannel.cs
@@ -37,7 +37,7 @@ namespace SS14.Shared.Network
         public string RemoteAddress => Connection.RemoteEndPoint.Address.ToString();
 
         /// <inheritdoc />
-        public int Ping => (int) Math.Round(Connection.AverageRoundtripTime * 1000);
+        public short Ping => (short) Math.Round(Connection.AverageRoundtripTime * 1000);
 
         /// <inheritdoc />
         public T CreateNetMessage<T>()
